requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
openai==1.6.0
azure-search-documents==11.4.0
azure-identity==1.15.0
redis==5.0.1
python-multipart==0.0.6
tiktoken==0.5.2
PyPDF2==3.0.1
python-dotenv==1.0.0
httpx==0.25.0
pydantic==2.5.0
